+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+ - The fourth basic collector, `Histogram`, have been added. ([#10][github-10])
 
 ## [1.1.0] - 2023-11-13
 ### Added
@@ -47,6 +49,7 @@ See [README.md](README.md) for details and documentation.
 
 [github-6]: https://github.com/fredrikekre/Prometheus.jl/pull/6
 [github-7]: https://github.com/fredrikekre/Prometheus.jl/pull/7
+[github-10]: https://github.com/fredrikekre/Prometheus.jl/pull/10
 
 [Unreleased]: https://github.com/fredrikekre/Prometheus.jl/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/fredrikekre/Prometheus.jl/compare/v1.0.1...v1.1.0